bill split settings are withing combo box
need to find a way to procedurally change the boxes depending on the
different options.

+++ mainwindow.cpp
@@ -115,7 +115,7 @@ void MainWindow::onUpdateTableInfo()
     Group* group = _instance->groupRecords.fetchLatestRecord();
 
     //PeopleHelper variables
-    int peopleCount = group->getPeopleCount();
+    size_t peopleCount = group->getPeopleCount();
     std::string personName = "";
 
     //go through the people records and get their names
